requirements.txt
uvicorn[standard]
motor
python-jose
passlib[bcrypt]
redis
fastapi
pydantic
pydantic-settings
pydantic[email]
aioredis
typing_extensions>=4.0.0
pytest
httpx
anyio
pytest-asyncio
pytest-tornasync
pytest-trio

